+++ CHANGES.md
@@ -17,7 +17,8 @@ Release Notes.
 * Add Seata in the component definition. Seata plugin hosts on Seata project.
 * Extended Kafka plugin to properly trace consumers that have topic partitions directly assigned.
 * Support print SkyWalking context to logs.
-* Add `MessageListener` enhancement in pulsar plugin
+* Add `MessageListener` enhancement in pulsar plugin.
+* fix a bug that spring-mvc set an error endpoint name if the controller class annotation implements an interface.
 * Add an optional agent plugin to support mybatis.
 
 #### OAP-Backend